pyproject.toml
[project]
name = "ResumeGenerator"
version = "0.1.0"
description = "This repo will work as my portfolio as well as will generate a pdf version of the resume."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
]
